+++ database/migrations/2016_04_06_172127_smarts_industrial_parameters.php
@@ -32,8 +32,8 @@ class SmartsIndustrialParameters extends Migration
             $table->string('site_facility_state');
             $table->string('site_facility_zip');
             $table->string('site_facility_county');
-            $table->decimal('latitude')->nullable();
-            $table->decimal('longitude')->nullable();
+            $table->double('latitude', 15, 10)->nullable();
+            $table->double('longitude', 15, 10)->nullable();
             $table->timestamp('date_time_of_sample_collection');
             $table->string('program_type');
             $table->string('primary_sic_1');
